chore(v0): remove @fluentui/react-northstar-styles-renderer
