+++ packages/fluentui/e2e/tsconfig.e2e.json
@@ -21,7 +21,6 @@
       "@fluentui/react-icons-northstar": ["packages/fluentui/react-icons-northstar/src/index.ts"],
       "@fluentui/react-northstar": ["packages/fluentui/react-northstar/src/index.ts"],
       "@fluentui/react-northstar-fela-renderer": ["packages/fluentui/react-northstar-fela-renderer/src/index.ts"],
-      "@fluentui/react-northstar-styles-renderer": ["packages/fluentui/react-northstar-styles-renderer/src/index.ts"],
       "@fluentui/react-northstar-prototypes": ["packages/fluentui/react-northstar-prototypes/src/index.ts"],
       "@fluentui/react-proptypes": ["packages/fluentui/react-proptypes/src/index.ts"],
       "@fluentui/styles": ["packages/fluentui/styles/src/index.ts"]
