+++ packages/fluentui/react-bindings/src/renderer/RendererContext.ts
@@ -1,5 +1,4 @@
-import { createFelaRenderer } from '@fluentui/react-northstar-fela-renderer';
-import { CreateRenderer } from '@fluentui/react-northstar-styles-renderer';
+import { createFelaRenderer, type CreateRenderer } from '@fluentui/react-northstar-fela-renderer';
 import * as React from 'react';
 
 export const RendererContext = React.createContext<CreateRenderer>(createFelaRenderer());
