+++ packages/fluentui/react-bindings/test/compose/useCompose-test.tsx
@@ -1,4 +1,4 @@
-import { noopRenderer } from '@fluentui/react-northstar-styles-renderer';
+import { noopRenderer } from '@fluentui/react-northstar-fela-renderer';
 import { ComponentSlotStylesPrepared, emptyTheme, ThemeInput } from '@fluentui/styles';
 import cx from 'classnames';
 import { mount, shallow } from 'enzyme';
