+++ packages/fluentui/react-bindings/tsconfig.json
@@ -30,9 +30,6 @@
     },
     {
       "path": "../react-northstar-fela-renderer"
-    },
-    {
-      "path": "../react-northstar-styles-renderer"
     }
   ]
 }
