+++ packages/fluentui/react-northstar-fela-renderer/package.json
@@ -5,7 +5,7 @@
   "bugs": "https://github.com/microsoft/fluentui/issues",
   "dependencies": {
     "@babel/runtime": "^7.10.4",
-    "@fluentui/react-northstar-styles-renderer": "^0.68.2",
+    "@fluentui/react-northstar-fela-renderer": "^0.68.2",
     "@fluentui/styles": "^0.68.2",
     "css-in-js-utils": "^3.0.0",
     "fela": "^10.6.1",
