+++ packages/fluentui/react-northstar-fela-renderer/src/createFelaRenderer.tsx
@@ -1,4 +1,3 @@
-import { CreateRenderer, Renderer } from '@fluentui/react-northstar-styles-renderer';
 import { createRenderer, IRenderer, IStyle, TPlugin } from 'fela';
 import felaPluginEmbedded from 'fela-plugin-embedded';
 import felaPluginFallbackValue from 'fela-plugin-fallback-value';
@@ -14,37 +13,7 @@ import { felaPerformanceEnhancer } from './felaPerformanceEnhancer';
 import { felaSanitizeCssPlugin } from './felaSanitizeCssPlugin';
 import { felaStylisEnhancer } from './felaStylisEnhancer';
 import { RendererProvider } from './RendererProvider';
-import { FelaRenderer, FelaRendererParam } from './types';
-
-let felaDevMode = false;
-
-try {
-  // eslint-disable-next-line no-undef
-  felaDevMode = !!window.localStorage.felaDevMode;
-} catch {}
-
-if (typeof window !== 'undefined' && process.env.NODE_ENV !== 'production' && process.env.NODE_ENV !== 'test') {
-  if (felaDevMode) {
-    /* eslint-disable-next-line no-console */
-    console.warn(
-      [
-        '@fluentui/react-northstar:',
-        'You are running Fela in development mode and this can cause performance degrades.',
-        'To disable it please paste `delete window.localStorage.felaDevMode` to your browsers console and reload current page.',
-      ].join(' '),
-    );
-  } else {
-    /* eslint-disable-next-line no-console */
-    console.warn(
-      [
-        '@fluentui/react-northstar:',
-        'You are running Fela in production mode.',
-        'This limits your ability to edit styles in browsers development tools.',
-        'To enable development mode please paste `window.localStorage.felaDevMode = true` to your browsers console and reload the page.',
-      ].join(' '),
-    );
-  }
-}
+import type { CreateRenderer, FelaRenderer, FelaRendererParam, Renderer } from './types';
 
 const blocklistedClassNames = [
   // Blocklist contains a list of classNames that are used by FontAwesome
@@ -68,7 +37,7 @@ const filterClassName = (className: string): boolean => {
 };
 
 const rendererConfig = {
-  devMode: felaDevMode,
+  devMode: false,
   filterClassName,
   enhancers: [felaPerformanceEnhancer, felaFocusVisibleEnhancer, felaStylisEnhancer],
   plugins: [
