+++ packages/fluentui/react-northstar-fela-renderer/src/index.ts
@@ -1,1 +1,4 @@
 export { createFelaRenderer } from './createFelaRenderer';
+export { noopRenderer } from './noopRenderer';
+
+export { type CreateRenderer, type RendererParam, type Renderer, type RendererRenderRule } from './types';
