+++ packages/fluentui/react-northstar-fela-renderer/src/noopRenderer.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Renderer } from './types';
+import type { Renderer } from './types';
 
 // Provides a minimal functionality to render components without styles and without runtime errors.
