+++ packages/fluentui/react-northstar-fela-renderer/tsconfig.json
@@ -19,7 +19,7 @@
   "include": ["src", "test"],
   "references": [
     {
-      "path": "../react-northstar-styles-renderer"
+      "path": "../styles"
     }
   ]
 }
