--- packages/fluentui/react-northstar-styles-renderer/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["plugin:@fluentui/eslint-plugin/react-northstar"],
-  "root": true
-}
