--- packages/fluentui/react-northstar-styles-renderer/.gulp.js
@@ -1,1 +1,0 @@
-module.exports = require('../../../.gulp');
