--- packages/fluentui/react-northstar-styles-renderer/babel.config.js
@@ -1,1 +1,0 @@
-module.exports = api => require('@fluentui/scripts-babel')(api);
