--- packages/fluentui/react-northstar-styles-renderer/gulpfile.ts
@@ -1,1 +1,0 @@
-import '../../../gulpfile';
