--- packages/fluentui/react-northstar-styles-renderer/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@fluentui/react-northstar-styles-renderer",
-  "description": "A noop renderer for FluentUI React Northstar with typings.",
-  "version": "0.68.2",
-  "bugs": "https://github.com/microsoft/fluentui/issues",
-  "dependencies": {
-    "@babel/runtime": "^7.10.4",
-    "@fluentui/styles": "^0.68.2"
-  },
-  "devDependencies": {
-    "@fluentui/eslint-plugin": "*",
-    "@fluentui/scripts-babel": "*",
-    "@fluentui/scripts-gulp": "*",
-    "@types/react": "17.0.44",
-    "@types/react-dom": "17.0.15",
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
-  },
-  "files": [
-    "dist"
-  ],
-  "homepage": "https://github.com/microsoft/fluentui/tree/master/packages/fluentui/react-northstar-styles-renderer",
-  "jsnext:main": "dist/es/index.js",
-  "license": "MIT",
-  "main": "dist/commonjs/index.js",
-  "module": "dist/es/index.js",
-  "peerDependencies": {
-    "react": "^18"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "microsoft/fluentui.git",
-  "scripts": {
-    "build": "gulp bundle:package:no-umd",
-    "clean": "gulp bundle:package:clean",
-    "lint": "ESLINT_USE_FLAT_CONFIG=false eslint --ext .js,.ts,.tsx .",
-    "lint:fix": "yarn lint --fix"
-  },
-  "sideEffects": false,
-  "types": "dist/es/index.d.ts"
-}
