--- packages/fluentui/react-northstar-styles-renderer/project.json
@@ -1,7 +1,0 @@
-{
-  "name": "react-northstar-styles-renderer",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "implicitDependencies": [],
-  "tags": ["react-northstar"]
-}
