--- packages/fluentui/react-northstar-styles-renderer/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.v0.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "dist/dts",
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noImplicitAny": true,
-    "noUnusedParameters": true,
-    "strictNullChecks": true
-  },
-  "include": ["src", "test"],
-  "references": [
-    {
-      "path": "../styles"
-    }
-  ]
-}
