+++ packages/fluentui/react-northstar/package.json
@@ -13,7 +13,7 @@
     "@fluentui/react-component-nesting-registry": "^0.68.2",
     "@fluentui/react-component-ref": "^0.68.2",
     "@fluentui/react-icons-northstar": "^0.68.2",
-    "@fluentui/react-northstar-styles-renderer": "^0.68.2",
+    "@fluentui/react-northstar-fela-renderer": "^0.68.2",
     "@fluentui/react-portal-compat-context": "^9.0.13",
     "@fluentui/react-proptypes": "^0.68.2",
     "@fluentui/styles": "^0.68.2",
