+++ packages/fluentui/react-northstar/test/utils/withProvider.tsx
@@ -1,5 +1,5 @@
 import { Unstable_FluentContextProvider, ProviderContextPrepared } from '@fluentui/react-bindings';
-import { Renderer, noopRenderer } from '@fluentui/react-northstar-styles-renderer';
+import { Renderer, noopRenderer } from '@fluentui/react-northstar-fela-renderer';
 import { emptyTheme, ThemePrepared } from '@fluentui/styles';
 import { mount, MountRendererProps, ComponentType } from 'enzyme';
 import * as React from 'react';
