+++ packages/fluentui/react-northstar/tsconfig.json
@@ -38,7 +38,7 @@
       "path": "../react-component-ref"
     },
     {
-      "path": "../react-northstar-styles-renderer"
+      "path": "../react-northstar-fela-renderer"
     },
     {
       "path": "../react-proptypes"
